Reorganize should_run tests into separate policy test file
